update(process): mix audio streams
re-written process command line to mix all audio streams
