+++ electron/helpers/ff/process.ts
@@ -1,6 +1,8 @@
 import {FFMpegProgress, IFFMpegProgressData} from "ffmpeg-progress-wrapper";
 import {FFHelpers} from "./index";
 import {AudioSettings, RenderStrategy, VideoSettings} from "../../types";
+import {MediaDetails} from "../../../common/ff/types";
+import {ff_getAudioTrackIndexes} from "../../../common/ff";
 
 export namespace VideoProcess {
 
@@ -61,14 +63,43 @@ export namespace VideoProcess {
     return filters;
   }
 
+  function filterComplex(mediaDetails: MediaDetails, videoSettings: VideoSettings, audioSettings: AudioSettings) {
+    const steps: string[] = [];
+    const mappings = new Set<string>();
+
+    steps.push(`[0:v]${settings2filters(videoSettings).join(',')}[v]`);
+    mappings.add('[v]')
+
+    const audioStreamIndexes = ff_getAudioTrackIndexes(mediaDetails);
+
+    if (audioStreamIndexes.length > 0 && audioSettings.volume > 0) {
+      const audioFilters = audio2filters(audioSettings);
+
+      const header = audioStreamIndexes.map((i) => `[0:${i}]`).join('');
+
+      if(audioStreamIndexes.length > 1){
+        audioFilters.unshift(`amix=${audioStreamIndexes.length}`);
+      }
+
+      steps.push(`${header}${audioFilters.join(',')}[a]`)
+      mappings.add('[a]');
+    }
+
+    return {
+      filter_complex: steps,
+      mappings: [...mappings]
+    };
+  }
+
   export function process(file: string,
-                          start: number,
-                          end: number,
-                          out: string,
-                          strategy: RenderStrategy,
-                          settings: VideoSettings,
-                          audio: AudioSettings,
-                          progress: (p: IFFMpegProgressData) => void
+    start: number,
+    end: number,
+    out: string,
+    strategy: RenderStrategy,
+    settings: VideoSettings,
+    audio: AudioSettings,
+    mediaDetails: MediaDetails,
+    progress: (p: IFFMpegProgressData) => void
   ) {
 
     const ffmpeg = FFHelpers.ffmpeg;
@@ -77,12 +108,19 @@ export namespace VideoProcess {
       out = `${file}.${start.toFixed()}-${end.toFixed()}.mp4`;
     }
 
+    const fc = filterComplex(mediaDetails, settings, audio);
+
+    console.log('filter_complex', fc.filter_complex);
+    console.log('filter_complex mappings', fc.mappings);
+
     const p = new FFMpegProgress([
         '-ss', start.toFixed(6),
         '-to', end.toFixed(6),
         '-i', file,
-        '-vf', settings2filters(settings).join(','),
-        '-af', audio2filters(audio).join(','),
+
+        ...(fc.filter_complex.length > 0 ? ['-filter_complex', fc.filter_complex.join(';')] : []),
+        ...(fc.mappings.length > 0 ? fc.mappings.map(x => ['-map', x]).flat() : []),
+
         ...strategy2params(strategy, end - start, audio?.volume > 0),
         ...(audio?.volume > 0 ? [] : ['-an']),
         '-c:v', 'libx264',
